update makefile
更新makefile定义

+++ Makefile
@@ -14,10 +14,11 @@ tar: # 打包
 	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o gormt main.go 
 	tar czvf gormt_linux.zip gormt config.yml
 clear:
-	- rm -rf  model/* 
-	- rm -rf err/ 
-	- rm gormt
-	- rm gormt.exe
-	- rm gormt_linux.zip
-	- rm gormt_mac.zip
-	- rm gormt_windows.zip
+	test ! -d model/ || rm -rf  model/*
+	test ! -d err/ || rm -rf  err/
+	test ! -f gormt || rm gormt
+	test ! -f gormt.exe || rm gormt.exe
+	test ! -f gormt_linux.zip || rm gormt_linux.zip
+	test ! -f gormt_mac.zip || rm gormt_mac.zip
+	test ! -f gormt_windows.zip || rm gormt_windows.zip
+
